Move test helper locally into the only test module where it's actually used/needed
