+++ packages/liveblocks-client/src/immutable.test.ts
@@ -4,18 +4,120 @@ import {
   createSerializedRegister,
   FIRST_POSITION,
   FOURTH_POSITION,
-  prepareStorageImmutableTest,
+  prepareRoomWithStorage,
   SECOND_POSITION,
+  serverMessage,
   THIRD_POSITION,
 } from "../test/utils";
 import { LiveList, LiveMap } from ".";
 import {
   legacy_patchImmutableObject,
+  lsonToJson,
   patchLiveObject,
   patchLiveObjectKey,
 } from "./immutable";
 import { LiveObject } from "./LiveObject";
-import type { JsonObject, StorageUpdate } from "./types";
+import type {
+  BaseUserMeta,
+  ClientMsg,
+  IdTuple,
+  Json,
+  JsonObject,
+  LsonObject,
+  SerializedCrdt,
+  StorageUpdate,
+  ToJson,
+} from "./types";
+import { ClientMsgCode, ServerMsgCode } from "./types";
+
+export async function prepareStorageImmutableTest<
+  TStorage extends LsonObject,
+  TPresence extends JsonObject = never,
+  TUserMeta extends BaseUserMeta = never,
+  TRoomEvent extends Json = never
+>(items: IdTuple<SerializedCrdt>[], actor: number = 0) {
+  let state = {} as ToJson<TStorage>;
+  let refState = {} as ToJson<TStorage>;
+
+  let totalStorageOps = 0;
+
+  const { machine: refMachine, storage: refStorage } =
+    await prepareRoomWithStorage<TPresence, TStorage, TUserMeta, TRoomEvent>(
+      items,
+      -1
+    );
+
+  const { machine, storage } = await prepareRoomWithStorage<
+    TPresence,
+    TStorage,
+    TUserMeta,
+    TRoomEvent
+  >(items, actor, (messages: ClientMsg<TPresence, TRoomEvent>[]) => {
+    for (const message of messages) {
+      if (message.type === ClientMsgCode.UPDATE_STORAGE) {
+        totalStorageOps += message.ops.length;
+        refMachine.onMessage(
+          serverMessage({
+            type: ServerMsgCode.UPDATE_STORAGE,
+            ops: message.ops,
+          })
+        );
+        machine.onMessage(
+          serverMessage({
+            type: ServerMsgCode.UPDATE_STORAGE,
+            ops: message.ops,
+          })
+        );
+      }
+    }
+  });
+
+  state = lsonToJson(storage.root) as ToJson<TStorage>;
+  refState = lsonToJson(refStorage.root) as ToJson<TStorage>;
+
+  const root = refStorage.root;
+  refMachine.subscribe(
+    root,
+    () => {
+      refState = lsonToJson(refStorage.root) as ToJson<TStorage>;
+    },
+    { isDeep: true }
+  );
+
+  function assert(
+    data: ToJson<TStorage>,
+    itemsCount?: number,
+    storageOpsCount?: number
+  ) {
+    assertStorage(data);
+
+    if (itemsCount) {
+      expect(machine.getItemsCount()).toBe(itemsCount);
+    }
+    expect(state).toEqual(refState);
+    expect(state).toEqual(data);
+
+    if (storageOpsCount) {
+      expect(totalStorageOps).toEqual(storageOpsCount);
+    }
+  }
+
+  function assertStorage(data: ToJson<TStorage>) {
+    const json = lsonToJson(storage.root);
+    expect(json).toEqual(data);
+    expect(lsonToJson(refStorage.root)).toEqual(data);
+  }
+
+  return {
+    storage,
+    refStorage,
+    assert,
+    assertStorage,
+    subscribe: machine.subscribe,
+    refSubscribe: refMachine.subscribe,
+    state,
+  };
+}
 
 function applyStateChanges<T extends JsonObject>(
   state: T,
