+++ packages/liveblocks-client/test/utils.ts
@@ -1,6 +1,5 @@
 import type { LiveObject } from "../src";
 import type { RoomAuthToken } from "../src/AuthToken";
-import { lsonToJson } from "../src/immutable";
 import { makePosition } from "../src/position";
 import type { Effects, Machine } from "../src/room";
 import { defaultState, makeStateMachine } from "../src/room";
@@ -21,8 +20,6 @@ import type {
   SerializedRootObject,
   ServerMsg,
   ToImmutable,
-  // XXX Should we deprecate this type in favor of ToImmutable?
-  ToJson,
 } from "../src/types";
 import { ClientMsgCode, CrdtType, ServerMsgCode } from "../src/types";
 import { remove } from "../src/utils";
@@ -155,7 +152,7 @@ const defaultContext = {
   },
 };
 
-async function prepareRoomWithStorage<
+export async function prepareRoomWithStorage<
   TPresence extends JsonObject,
   TStorage extends LsonObject,
   TUserMeta extends BaseUserMeta,
@@ -470,95 +467,6 @@ export async function reconnect<
       items: newItems,
     })
   );
-}
-
-export async function prepareStorageImmutableTest<
-  TStorage extends LsonObject,
-  TPresence extends JsonObject = never,
-  TUserMeta extends BaseUserMeta = never,
-  TRoomEvent extends Json = never
->(items: IdTuple<SerializedCrdt>[], actor: number = 0) {
-  let state = {} as ToJson<TStorage>;
-  let refState = {} as ToJson<TStorage>;
-
-  let totalStorageOps = 0;
-
-  const { machine: refMachine, storage: refStorage } =
-    await prepareRoomWithStorage<TPresence, TStorage, TUserMeta, TRoomEvent>(
-      items,
-      -1
-    );
-
-  const { machine, storage } = await prepareRoomWithStorage<
-    TPresence,
-    TStorage,
-    TUserMeta,
-    TRoomEvent
-  >(items, actor, (messages: ClientMsg<TPresence, TRoomEvent>[]) => {
-    for (const message of messages) {
-      if (message.type === ClientMsgCode.UPDATE_STORAGE) {
-        totalStorageOps += message.ops.length;
-        refMachine.onMessage(
-          serverMessage({
-            type: ServerMsgCode.UPDATE_STORAGE,
-            ops: message.ops,
-          })
-        );
-        machine.onMessage(
-          serverMessage({
-            type: ServerMsgCode.UPDATE_STORAGE,
-            ops: message.ops,
-          })
-        );
-      }
-    }
-  });
-
-  state = lsonToJson(storage.root) as ToJson<TStorage>;
-  refState = lsonToJson(refStorage.root) as ToJson<TStorage>;
-
-  const root = refStorage.root;
-  refMachine.subscribe(
-    root,
-    () => {
-      refState = lsonToJson(refStorage.root) as ToJson<TStorage>;
-    },
-    { isDeep: true }
-  );
-
-  function assert(
-    data: ToJson<TStorage>,
-    itemsCount?: number,
-    storageOpsCount?: number
-  ) {
-    assertStorage(data);
-
-    if (itemsCount) {
-      expect(machine.getItemsCount()).toBe(itemsCount);
-    }
-    expect(state).toEqual(refState);
-    expect(state).toEqual(data);
-
-    if (storageOpsCount) {
-      expect(totalStorageOps).toEqual(storageOpsCount);
-    }
-  }
-
-  function assertStorage(data: ToJson<TStorage>) {
-    const json = lsonToJson(storage.root);
-    expect(json).toEqual(data);
-    expect(lsonToJson(refStorage.root)).toEqual(data);
-  }
-
-  return {
-    storage,
-    refStorage,
-    assert,
-    assertStorage,
-    subscribe: machine.subscribe,
-    refSubscribe: refMachine.subscribe,
-    state,
-  };
 }
 
 export function createSerializedObject(
